move reset game state to world actions
