+++ src/features/dialog-manager/dialogs/endless-game-start/index.js
@@ -7,7 +7,7 @@ import store            from '../../../../config/store';
 import generateMap      from '../../../../modules/generate-map';
 import generateMonsters from '../../../../modules/generate-monsters';
 import exploreTiles     from '../../../../features/player/actions/explore-tiles';
-import { uuidv4 }       from '../../../../modules/uuid-v4';
+import uuidv4           from '../../../../modules/uuid-v4';
 
 import './styles.scss';
