+++ src/features/world/actions/reset-game-state.js
@@ -1,0 +1,17 @@
+
+export default function resetGameState() {
+  return dispatch => {
+    dispatch({
+      type: 'RESET',
+      payload: {}
+    });
+
+    dispatch({
+      type: 'PAUSE',
+      payload: {
+        pause: true,
+        gameStart: true
+      }
+    });
+  }
+}
