--- src/modules/reset-game-state.js
@@ -1,17 +1,0 @@
-import store from '../config/store';
-
-export default function resetGameState() {
-
-  store.dispatch({
-    type: 'RESET',
-    payload: {}
-  });
-
-  store.dispatch({
-    type: 'PAUSE',
-    payload: {
-      pause: true,
-      gameStart: true
-    }
-  });
-}
